move tests to separate module
